refactor : CoinRepository, ItemRepository를 CoinService, ItemService로 변경
