+++ src/main/java/vendingmachine/service/ItemService.java
@@ -1,15 +1,15 @@
-package vendingmachine.repository;
+package vendingmachine.service;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import vendingmachine.domain.Item;
 
-public class ItemRepository {
+public class ItemService {
 
 	private static final List<Item> itemRepository = new ArrayList<>();
 
-	public void save(Item item) {
+	public void register(Item item) {
 		itemRepository.add(item);
 	}
 }
